webhook-repo/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.13.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
certifi==2025.6.15
